fix: portability for macOS <execution> and GCC warnings
- Add portability_execution.hpp with __cpp_lib_execution detection;
  provides fallback sequential/parallel policy types when <execution>
  is unavailable (Apple libc++, older libstdc++)
- Replace all #include <execution> with portability header
- Fix GCC -Wuseless-cast by using uint64_t loop variable in
  test_art_bulk_load.cpp and DISABLE_GCC_WARNING in structural test
- Fix GCC -Wdangling-else by adding braces around EXPECT_GE/EXPECT_GT
  after bare if in test_art_bulk_load_structural.cpp

+++ test/test_art_bulk_load.cpp
@@ -38,13 +38,13 @@ constexpr auto val = value_view{val_bytes};
     std::size_t n) {
   std::vector<std::pair<std::uint64_t, value_view>> kv;
   kv.reserve(n);
-  for (std::size_t i = 0; i < n && i < 256; ++i) {
-    const auto key = static_cast<std::uint64_t>(i) << 56U;
+  for (std::uint64_t i = 0; i < n && i < 256; ++i) {
+    const auto key = i << 56U;
     kv.emplace_back(key, val);
   }
   // Overflow keys: share byte 0 == 0x00, differ at byte 1
-  for (std::size_t i = 256; i < n; ++i) {
-    const auto key = static_cast<std::uint64_t>(i - 255) << 48U;
+  for (std::uint64_t i = 256; i < n; ++i) {
+    const auto key = (i - 255) << 48U;
     kv.emplace_back(key, val);
   }
   std::ranges::sort(kv, {}, &decltype(kv)::value_type::first);
